TradingView_Chart_Python/requirements.txt
lightweight-charts
yfinance
pandas_ta
ib_insync
nest_asyncio